backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
slowapi==0.1.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
cryptography>=41.0.0

# HTTP Clients
httpx~=0.25.2
aiohttp==3.9.1
requests==2.31.0

# Telegram
python-telegram-bot==20.7

# Data Processing
pandas==2.1.4
numpy==1.26.3
pydantic==2.5.3
pydantic-settings==2.1.0

# PDF Generation
reportlab==4.0.9
PyPDF2==3.0.1
fpdf2==2.7.7

# AI/ML
openai==1.10.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Monitoring & Logging
loguru==0.7.2
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx~=0.25.2

# Utils
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0
